Move is_braille? and helper method to Identifiable module and include in Translator and Formatter
